+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - FEATURE: mid param router ex /users/:id/name by @poteto0 in #122
 - REFACTOR: some switch case by @poteto0 in #122
+- FEATURE: ctx.DebugParam by @poteto0 in #125
 
 ## 0.23.x